+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Fix bug where color for retained messages is lost
 * Fix bug where `<host>:<port>` was not allowed
 * Add logging to `~/.cache/mqttui/mqttui.log`
+* Add message history
+* Allow scrolling with '{' & '}' like in VIM
 
 ## 1.1.0